Move tokensTransferred action to the token module
